requirements.txt
praw
langchain
google-generativeai
ipython
python-dotenv